+++ lib/joken.ex
@@ -16,7 +16,11 @@ defmodule Joken do
 
       defmodule My.Parameters.Module do
         alias Poison, as: JSON
+<<<<<<< HEAD
         @behaviour Joken.Parameters
+=======
+        @behaviour Joken.Codec
+>>>>>>> master
 
         def encode(map) do
           JSON.encode!(map)
@@ -27,7 +31,7 @@ defmodule Joken do
         end
       end
 
-       config :joken
+       config :joken,
          secret_key: "test",
          parameters_module: My.Parameters.Module,
          algorithm: :HS256, #Optional. defaults to :HS256