+++ test/joken_token_test.exs
@@ -67,6 +67,7 @@ defmodule Joken.Token.Test do
     assert {:ok, _} = Joken.Token.decode(@secret, @poison_json_module, @unsorted_payload_token)
   end
 
+
   test "error with invalid algorithm" do
     {:error, message} = Joken.Token.encode(@secret, @poison_json_module, @payload, :HS1024)
     assert message == "Unsupported algorithm"
@@ -94,6 +95,20 @@ defmodule Joken.Token.Test do
 
     {:ok, decoded_payload} = Joken.Token.decode(@secret, @jsx_json_module, token, :HS512) 
     assert(@payload == decoded_payload) 
+  end
+
+  test "missing signature" do
+    {:ok, token} = Joken.Token.encode(@secret, @jsx_json_module, @payload, :HS512)
+    assert(token == "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSm9obiBEb2UifQ.zi1zohSNwRdHftnWKE16vE3VmbGFtG27LxbYDXAodVlX7T3ATgmJJPjluwf2SPKJND2-O7alOq8NWv6EAnWWyg")
+
+    {:error, message} = Joken.Token.decode(@secret, @jsx_json_module, "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSm9obiBEb2UifQ", :HS512) 
+    assert("Missing signature" == message) 
+  end
+
+  test "unsecure token" do
+    {status, message} = Joken.Token.decode(@secret, @jsx_json_module, "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSm9obiBEb2UifQ", :none)
+    assert(status == :error)
+    assert(message == "Missing signature")
   end
 
   test "decode token generated with un-sorted keys (JSX)" do