+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 # v0.14.0-dev
 
 * Enhancements
@@ -7,6 +8,11 @@
 * Breaking
   * `Joken.Codec` is replaced by `Joken.Parameters`. 
   * `json_module` in config is replaced by `parameters_module`. 
+=======
+# v0.13.1
+  * Enhancements
+    * Checking to make sure signature is on token unless `:none` is passed as the algorithm
+>>>>>>> master
 
 # v0.13.0